% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorreg_Wrapper.R
\name{getmean}
\alias{getmean}
\title{posterior mean for tensor regression}
\usage{
getmean(X, beta, rank, rank.exclude = NULL)
}
\arguments{
\item{X}{Input tensor.}

\item{beta}{Estimated coefficient tensor.}

\item{rank}{Rank used in tensor decomposition.}

\item{rank.exclude}{Optional rank values to exclude.}
}
\value{
Mean tensor.
}
\description{
posterior mean for tensor regression
}
