% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmean_cpp.R
\name{getmean_cpp}
\alias{getmean_cpp}
\title{Posterior Mean Using C++}
\usage{
getmean_cpp(X_vec, beta, n, p, d, rank)
}
\arguments{
\item{X_vec}{Flattened tensor (numeric vector of length n*p*d)}

\item{beta}{List of p×d matrices (length = rank)}

\item{n}{Number of observations}

\item{p}{Number of rows in each beta matrix}

\item{d}{Number of columns in each beta matrix}

\item{rank}{Rank of tensor decomposition}
}
\value{
Numeric vector of length n
}
\description{
Calls `getmean_cpp` to compute the posterior mean from tensor X and beta list.
}
