% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorreg_Wrapper.R
\name{predict.tensor.reg}
\alias{predict.tensor.reg}
\title{Prediction from tensor regression (S3 method)}
\usage{
\method{predict}{tensor.reg}(object, x.new, z.new, scale = TRUE, ...)
}
\arguments{
\item{object}{tensor.reg object}

\item{x.new}{new tensor predictors (n x p x d)}

\item{z.new}{new scalar covariates (n x pgamma)}

\item{scale}{Logical; whether to scale predictors}

\item{...}{additional arguments}
}
\value{
Predicted response vector
}
\description{
Prediction from tensor regression (S3 method)
}
