% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorreg_Wrapper.R
\name{predict_tensor_reg}
\alias{predict_tensor_reg}
\title{Predict tensor regression (wrapper)}
\usage{
predict_tensor_reg(fit, x.new, z.new, scale = TRUE)
}
\arguments{
\item{fit}{tensor.reg object}

\item{x.new}{new tensor predictors}

\item{z.new}{new scalar covariates}

\item{scale}{Logical; whether to scale predictors}
}
\value{
Predicted response vector
}
\description{
Predict tensor regression (wrapper)
}
