% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorreg_Wrapper.R
\name{rigamma}
\alias{rigamma}
\title{Inverse-gamma random number generator}
\usage{
rigamma(n, shape, rate)
}
\arguments{
\item{n}{Number of samples.}

\item{shape}{Shape parameter of the gamma distribution.}

\item{rate}{Rate parameter of the gamma distribution.}
}
\value{
A numeric vector of inverse-gamma samples.
}
\description{
Inverse-gamma random number generator
}
