% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorreg_Wrapper.R
\name{rmse}
\alias{rmse}
\title{root-mean-square error (RMSE)}
\usage{
rmse(a, b)
}
\arguments{
\item{a}{Predicted values.}

\item{b}{True observed values.}
}
\value{
RMSE value.
}
\description{
root-mean-square error (RMSE)
}

