% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensorreg_Wrapper.R
\name{cv.tensor.reg}
\alias{cv.tensor.reg}
\title{Cross-validation for tensor regression}
\usage{
cv.tensor.reg(x.train, z.train, y.train, ranks = 1:3, nsweep = 50)
}
\arguments{
\item{x.train}{Training tensor X data.}

\item{z.train}{Training scalar covariates.}

\item{y.train}{Training response vector.}

\item{ranks}{A vector of rank values to test.}

\item{nsweep}{Number of MCMC sweeps / iterations.}
}
\value{
A data.frame containing the best rank and RMSE performance.
}
\description{
Cross-validation for tensor regression
}
